Serve the theme preview page from the shared styles route
Registers the styles route on both site-editor-v2 and theme-preview
pages instead of duplicating its stage and canvas, and gates the preview
additions — the beforeLoad guard, the "Previewing X" title, and the
Activate button with v1's review-and-activate panel — on the
wp_theme_preview parameter. The preview page URL gains p=/styles.
