+++ routes/styles/package.json
@@ -4,10 +4,14 @@
 	"private": true,
 	"route": {
 		"path": "/styles",
-		"page": "site-editor-v2"
+		"page": [
+			"site-editor-v2",
+			"theme-preview"
+		]
 	},
 	"dependencies": {
 		"@wordpress/admin-ui": "file:../../packages/admin-ui",
+		"@wordpress/api-fetch": "file:../../packages/api-fetch",
 		"@wordpress/base-styles": "file:../../packages/base-styles",
 		"@wordpress/components": "file:../../packages/components",
 		"@wordpress/compose": "file:../../packages/compose",
@@ -15,6 +19,7 @@
 		"@wordpress/data": "file:../../packages/data",
 		"@wordpress/editor": "file:../../packages/editor",
 		"@wordpress/element": "file:../../packages/element",
+		"@wordpress/html-entities": "file:../../packages/html-entities",
 		"@wordpress/i18n": "file:../../packages/i18n",
 		"@wordpress/icons": "file:../../packages/icons",
 		"@wordpress/lazy-editor": "file:../../packages/lazy-editor",
