+++ routes/styles/route.ts
@@ -1,6 +1,13 @@
 import { resolveSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
+import { decodeEntities } from '@wordpress/html-entities';
 import { __ } from '@wordpress/i18n';
+import { notFound } from '@wordpress/route';
+import {
+	getPreviewedStylesheet,
+	getPreviewTitle,
+	isThemePreviewAdminPage,
+} from './previewed-theme';
 
 async function isBlockTheme() {
 	const currentTheme = await resolveSelect( coreStore ).getCurrentTheme();
@@ -9,9 +16,30 @@ async function isBlockTheme() {
 
 /**
  * Route configuration for styles.
+ *
+ * The route also serves the theme preview page, rendering the same screen
+ * for the theme named by the `wp_theme_preview` query parameter, which Core
+ * resolves on every request — REST requests included. See
+ * `gutenberg_get_theme_preview_url()` for the full account.
  */
 export const route = {
-	title: () => __( 'Styles' ),
+	beforeLoad: () => {
+		if ( isThemePreviewAdminPage() && ! getPreviewedStylesheet() ) {
+			throw notFound();
+		}
+	},
+
+	title: async () => {
+		if ( ! getPreviewedStylesheet() ) {
+			return __( 'Styles' );
+		}
+		const theme = await resolveSelect( coreStore ).getCurrentTheme();
+		return getPreviewTitle(
+			theme?.name?.rendered
+				? decodeEntities( theme.name.rendered )
+				: undefined
+		);
+	},
 
 	// Classic themes have no global styles to edit, so the style book is all
 	// this route has to show.
