--- routes/theme-preview/package.json
@@ -1,29 +1,0 @@
-{
-	"name": "@wordpress/theme-preview-route",
-	"version": "1.0.0",
-	"private": true,
-	"route": {
-		"path": "/",
-		"page": [
-			"theme-preview"
-		]
-	},
-	"dependencies": {
-		"@wordpress/admin-ui": "file:../../packages/admin-ui",
-		"@wordpress/api-fetch": "file:../../packages/api-fetch",
-		"@wordpress/components": "file:../../packages/components",
-		"@wordpress/compose": "file:../../packages/compose",
-		"@wordpress/core-data": "file:../../packages/core-data",
-		"@wordpress/data": "file:../../packages/data",
-		"@wordpress/editor": "file:../../packages/editor",
-		"@wordpress/element": "file:../../packages/element",
-		"@wordpress/html-entities": "file:../../packages/html-entities",
-		"@wordpress/i18n": "file:../../packages/i18n",
-		"@wordpress/icons": "file:../../packages/icons",
-		"@wordpress/lazy-editor": "file:../../packages/lazy-editor",
-		"@wordpress/route": "file:../../packages/route",
-		"@wordpress/routes-lock-unlock": "file:../lock-unlock",
-		"@wordpress/ui": "file:../../packages/ui",
-		"@wordpress/url": "file:../../packages/url"
-	}
-}
