--- routes/theme-preview/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"$schema": "https://json.schemastore.org/tsconfig.json",
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"jsx": "react-jsx",
-		"rootDir": ".",
-		"noEmit": true,
-		"emitDeclarationOnly": false,
-		"composite": false,
-		"types": [ "style-imports" ]
-	},
-	"include": [ "**/*.ts", "**/*.tsx" ],
-	"exclude": [ "build", "node_modules" ]
-}
